# internal/web/static/openapi.json
{
    "swagger": "2.0",
    "info": {
        "description": "The API the Tunnel Manager screens are built on. Every path\nbelow sits under /api and needs a session, except POST\n/api/login and GET /api/setup.\n\nLogging in sets two cookies. Over HTTPS they are named\n__Host-tm_session and __Host-tm_csrf, and over plain HTTP they\nare tm_session and tm_csrf. A browser sends them back on its\nown, so \"Try it out\" on this page works as soon as you are\nlogged in to the screens in the same browser.\n\nEvery POST, PUT and DELETE also needs the token of the session\nin an X-CSRF-Token header. POST /api/login answers with it in\ndata.csrf_token; paste it into Authorize above and this page\nsends it. A GET needs no token.\n\nEvery answer has the same shape: {\"success\":true,\"data\":...}\nor {\"success\":false,\"error\":\"...\"}. An answer that says no\ncarries error_code, which names the refusal and is what a\nscript should match on, and error_args, the values its English\nsentence was written with.",
        "title": "Tunnel Manager API",
        "contact": {},
        "license": {
            "name": "Apache 2.0"
        },
        "version": "3.8.3"
    },
    "basePath": "/api",
    "paths": {
...
        "/update/check": {
            "post": {
                "security": [
                    {
                        "CSRFToken": []
                    }
                ],
                "description": "Answers what GET /api/update would then answer. A POST because it makes a request to another host, which is not a thing a link or a prefetch may set off.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "update"
                ],
                "summary": "Read the newest release now",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api.updateView"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/update/install": {
            "post": {
                "security": [
                    {
                        "CSRFToken": []
                    }
                ],
                "description": "Takes the account password. It answers that the install started and never that it finished: what the install ends with is a restart of the service answering the request.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "update"
                ],
                "summary": "Start the install of the newest release",
                "parameters": [
                    {
                        "description": "The account password",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.updateInstallRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api.updateView"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "The password does not open this account",
                        "schema": {
                            "$ref": "#/definitions/api.errorBody"
                        }
                    },
                    "409": {
                        "description": "There is nothing newer, or this build cannot install one",
                        "schema": {
                            "$ref": "#/definitions/api.errorBody"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "api.errorArgs": {
            "type": "object",
            "additionalProperties": {
                "type": "string"
            }
        },
        "api.errorBody": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "error_args": {
                    "$ref": "#/definitions/api.errorArgs"
                },
                "error_code": {
                    "$ref": "#/definitions/api.errorCode"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "api.errorCode": {
            "type": "string",
            "enum": [
                "unspecified",
                "request.body.invalid",
                "request.validation.failed",
                "db.transaction.begin_failed",
                "db.transaction.commit_failed",
                "list.page.not_a_number",
                "list.size.unsupported",
                "auth.session.required",
                "auth.csrf.refused",
                "auth.credentials.invalid",
                "auth.session.create_failed",
                "auth.setup.required",
                "auth.setup.already_done",
                "auth.setup.failed",
                "auth.password.too_short",
                "auth.password.too_long",
                "auth.attempts.too_many",
                "auth.password_attempts.too_many",
                "account.read_failed",
                "account.store_failed",
                "account.password.hash_failed",
                "account.username.empty",
                "account.request.invalid",
                "account.nothing_to_change",
                "account.current_password.wrong",
                "account.username.unchanged",
                "account.new_password.unchanged",
                "account.new_password.too_short",
                "account.new_password.too_long",
                "host.id.invalid",
                "host.not_found",
                "host.fetch_failed",
                "host.list_failed",
                "host.create_failed",
                "host.update_failed",
                "host.delete_failed",
                "host.password.encrypt_failed",
                "host.private_key.encrypt_failed",
                "host.create.no_login",
                "host.create.key_refused",
                "host.create.service_ports_read_failed",
                "host.create.service_ports_store_failed",
                "host.update.key_refused",
                "host.update.passphrase_without_key",
                "host.host_key.nothing_to_approve",
                "host.host_key.fingerprint_changed",
                "host.host_key.password_wrong",
                "service_port.id.invalid",
                "service_port.not_found",
                "service_port.fetch_failed",
                "service_port.list_failed",
                "service_port.create_failed",
                "service_port.update_failed",
                "service_port.delete_failed",
                "service_port.create.hosts_read_failed",
                "service_port.create.hosts_store_failed",
                "assignment.add_and_remove",
                "assignment.service_port.not_found",
                "assignment.update_failed",
                "status.desired_count_failed",
                "status.fetch_failed",
                "settings.read_failed",
                "settings.store_failed",
                "settings.refused",
                "settings.ui_language.unsupported",
                "certificate.https_off",
                "certificate.served.read_failed",
                "certificate.renew_failed",
                "certificate.install.refused",
                "certificate.store_failed",
                "certificate.read_back_failed",
                "logs.lines.not_a_number",
                "logs.lines.below_one",
                "logs.file.not_configured",
                "logs.file.missing",
                "logs.file.read_failed",
                "logs.clear.password_missing",
                "logs.clear.password_wrong",
                "logs.clear.failed",
                "update.check_failed",
                "update.not_installable",
                "update.password_missing",
                "update.password_wrong",
                "update.install_failed",
                "export.password.required",
                "export.password.too_short",
                "export.password.too_long",
                "export.hosts.read_failed",
                "export.service_ports.read_failed",
                "export.assignments.read_failed",
                "export.host.secrets_unreadable",
                "export.seal_failed",
                "import.file.missing",
                "import.password.required",
                "import.file.not_an_export",
                "import.file.damaged",
                "import.password.wrong",
                "import.file.open_failed",
                "import.file.not_ours",
                "import.file.wrong_kind",
                "import.file.newer_format",
                "import.file.newer_format_by",
                "import.file.empty",
                "import.tunnels.unreadable",
                "import.host.refused",
                "import.host.no_login",
                "import.hosts.read_failed",
                "import.host.password_encrypt_failed",
                "import.host.key_refused",
                "import.host.private_key_encrypt_failed",
                "import.host.replace_failed",
                "import.host.create_failed",
                "import.service_port.refused",
                "import.service_ports.read_failed",
                "import.service_port.create_failed",
                "import.service_port.two_rows",
                "import.service_port.replace_failed",
                "import.assignments.host_read_failed",
                "import.assignments.clear_failed",
                "import.assignments.store_failed",
                "import.settings.unreadable",
                "import.settings.refused",
                "uninstall.password.wrong"
            ],
            "x-enum-varnames": [
                "errorCodeUnspecified",
                "errRequestBodyInvalid",
                "errRequestValidationFailed",
                "errTransactionBeginFailed",
                "errTransactionCommitFailed",
                "errListPageNotANumber",
                "errListSizeUnsupported",
                "errAuthRequired",
                "errAuthCSRFRefused",
                "errAuthCredentialsInvalid",
                "errAuthSessionCreateFailed",
                "errAuthSetupRequired",
                "errAuthSetupAlreadyDone",
                "errAuthSetupFailed",
                "errAuthPasswordTooShort",
                "errAuthPasswordTooLong",
                "errAuthTooManyAttempts",
                "errAuthPasswordTooManyAttempts",
                "errAccountReadFailed",
                "errAccountStoreFailed",
                "errAccountHashFailed",
                "errAccountUsernameEmpty",
                "errAccountRequestInvalid",
                "errAccountNothingToChange",
                "errAccountPasswordWrong",
                "errAccountUsernameUnchanged",
                "errAccountPasswordUnchanged",
                "errAccountNewPasswordShort",
                "errAccountNewPasswordLong",
                "errHostIDInvalid",
                "errHostNotFound",
                "errHostFetchFailed",
                "errHostListFailed",
                "errHostCreateFailed",
                "errHostUpdateFailed",
                "errHostDeleteFailed",
                "errHostPasswordEncryptFailed",
                "errHostKeyEncryptFailed",
                "errHostCreateNoLogin",
                "errHostCreateKeyRefused",
                "errHostCreateServicePortsRead",
                "errHostCreateServicePortsStore",
                "errHostUpdateKeyRefused",
                "errHostUpdatePassphraseAlone",
                "errHostKeyNothingToApprove",
                "errHostKeyFingerprintChanged",
                "errHostKeyPasswordWrong",
                "errServicePortIDInvalid",
                "errServicePortNotFound",
                "errServicePortFetchFailed",
                "errServicePortListFailed",
                "errServicePortCreateFailed",
                "errServicePortUpdateFailed",
                "errServicePortDeleteFailed",
                "errServicePortCreateHostsRead",
                "errServicePortCreateHostsStore",
                "errAssignmentAddAndRemove",
                "errAssignmentServicePortsMissing",
                "errAssignmentUpdateFailed",
                "errStatusDesiredCountFailed",
                "errStatusFetchFailed",
                "errSettingsReadFailed",
                "errSettingsStoreFailed",
                "errSettingsRefused",
                "errSettingsLanguageUnsupported",
                "errCertificateHTTPSOff",
                "errCertificateServedUnread",
                "errCertificateRenewFailed",
                "errCertificateInstallRefused",
                "errCertificateStoreFailed",
                "errCertificateReadBackFailed",
                "errLogsLinesNotANumber",
                "errLogsLinesBelowOne",
                "errLogsFileNotConfigured",
                "errLogsFileMissing",
                "errLogsFileReadFailed",
                "errLogsClearPasswordMissing",
                "errLogsClearPasswordWrong",
                "errLogsClearFailed",
                "errUpdateCheckFailed",
                "errUpdateNotInstallable",
                "errUpdatePasswordMissing",
                "errUpdatePasswordWrong",
                "errUpdateInstallFailed",
                "errExportPasswordRequired",
                "errExportPasswordTooShort",
                "errExportPasswordTooLong",
                "errExportHostsReadFailed",
                "errExportServicePortsRead",
                "errExportAssignmentsRead",
                "errExportHostSecretsSealed",
                "errExportSealFailed",
                "errImportFileMissing",
                "errImportPasswordRequired",
                "errImportFileNotAnExport",
                "errImportFileDamaged",
                "errImportPasswordWrong",
                "errImportFileOpenFailed",
                "errImportFileNotOurs",
                "errImportFileWrongKind",
                "errImportFileNewerFormat",
                "errImportFileNewerFormatBy",
                "errImportFileEmpty",
                "errImportTunnelsUnreadable",
                "errImportHostRefused",
                "errImportHostNoLogin",
                "errImportHostsReadFailed",
                "errImportHostPasswordEncrypt",
                "errImportHostKeyRefused",
                "errImportHostKeyEncrypt",
                "errImportHostReplaceFailed",
                "errImportHostCreateFailed",
                "errImportServicePortRefused",
                "errImportServicePortsReadFailed",
                "errImportServicePortCreate",
                "errImportServicePortTwoRows",
                "errImportServicePortReplace",
                "errImportAssignmentsHostRead",
                "errImportAssignmentsClearFailed",
                "errImportAssignmentsStoreFailed",
                "errImportSettingsUnreadable",
                "errImportSettingsRefused",
                "errUninstallPasswordWrong"
            ]
        },
...
        "api.updateInstallRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                }
            }
        },
        "api.updateView": {
            "type": "object",
            "properties": {
                "checked_at": {
                    "description": "Checked is when the last look happened, and the three fields under it are\nwhat it found. Checked is the zero time where nothing has looked yet,\nwhich the screen draws as never rather than as 1970.",
                    "type": "string"
                },
                "comparable": {
                    "type": "boolean"
                },
                "installable": {
                    "description": "Installable says whether the press that installs is offered. See the\nfield of the same name on the handler.",
                    "type": "boolean"
                },
                "newer": {
                    "type": "boolean"
                },
                "problem": {
                    "description": "Problem is why the last look found nothing, empty where it worked. A\ncheck that failed is not a version that is up to date, and this is what\nkeeps the two apart on the screen.",
                    "type": "string"
                },
                "tag": {
                    "type": "string"
                },
                "version": {
                    "description": "Version is what this process is running, without the leading v, the way\nthe version constant carries it.",
                    "type": "string"
                }
            }
        },
...
        "models.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "CSRFToken": {
            "description": "The token POST /api/login answers with, in\ndata.csrf_token. Required on every POST, PUT\nand DELETE. The session cookie goes with it and\nis sent by the browser on its own.",
            "type": "apiKey",
            "name": "X-CSRF-Token",
            "in": "header"
        }
    }
}